+++ Android. Module 10. Async operations/app/src/main/java/com/example/asyncoperations/model/room/entities/Episode.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class Episode(
     @PrimaryKey
-    val idEpisode: Int,
+    val idEpisode: Long,
     val airDate: String?,
     val episode: String?,
     val name: String?,
