+++ Android. Module 10. Async operations/app/src/main/java/com/example/asyncoperations/model/ui/CharacterUI.kt
@@ -2,7 +2,7 @@ package com.example.asyncoperations.model.ui
 
 
 data class CharacterUI(
-    val idCharacter: Int,
+    val idCharacter: Long,
     val image: String?,
     val name: String?,
     val species: String?,
