+++ Android. Module 10. Async operations/app/src/main/java/com/example/asyncoperations/model/ui/EpisodeUI.kt
@@ -1,7 +1,7 @@
 package com.example.asyncoperations.model.ui
 
 data class EpisodeUI(
-    val idEpisode: Int,
+    val idEpisode: Long,
     val airDate: String?,
     val episode: String?,
     val name: String?,
